requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-WTF==1.2.1
Werkzeug==2.3.7
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Pillow==10.0.1
torch==2.1.0
torchvision==0.16.0
matplotlib==3.7.3
numpy==1.24.4
xhtml2pdf==0.2.10
reportlab==4.0.4
python-dateutil==2.8.2
PyMySQL==1.1.0
pdfkit==1.0.0
Jinja2==3.1.2
Werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0.post2
gunicorn==21.2.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.0.5
